requirements.txt
absl-py==0.8.0
alabaster==0.7.12
amqp==2.5.0
# anaconda-client==1.7.2
# anaconda-navigator==1.9.7
# anaconda-project==0.8.2
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
astropy==3.1.2
atomicwrites==1.3.0
attrs==19.1.0
auto-tagify==1.4
autopep8==1.4.4
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
billiard==3.6.0.0
bitarray==0.8.3
bkcharts==0.2
bleach==3.1.0
bokeh==1.0.4
boto==2.49.0
boto3==1.9.175
botocore==1.12.175
Bottleneck==1.2.1
bs4==0.0.1
CacheControl==0.12.5
cachetools==3.1.1
camera==1.3.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
clyent==1.2.2
colorama==0.4.1
conda==4.6.11
conda-build==3.17.8
conda-verify==3.1.1
contextlib2==0.5.5
crypto==1.4.1
cryptography==2.6.1
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.6
cytoolz==0.9.0.1
dash==1.2.0
dash-core-components==1.1.2
dash-html-components==1.0.1
dash-renderer==1.0.1
dash-table==4.2.0
dask==1.1.4
decorator==4.4.0
defusedxml==0.5.0
dill==0.3.0
distributed==1.26.0
Django==2.2.2
dnspython==1.16.0
docutils==0.14
entrypoints==0.3
enum34==1.1.6
et-xmlfile==1.0.1
eventlet==0.25.1
fake-useragent==0.1.11
fastcache==1.0.2
filelock==3.0.10
firebase-admin==3.0.0
Flask==1.1.1
Flask-Compress==1.4.0
Flask-SocketIO==2.8.6
future==0.17.1
gast==0.2.2
gcloud==0.17.0
gdown==3.8.2
gensim==3.7.3
gevent==1.4.0
Glances==3.1.1
glob2==0.6
gmpy2==2.0.8
google==2.0.2
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.4.2
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.19.1
google-pasta==0.1.7
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
googletrans==2.4.0
graphviz==0.11.1
greenlet==0.4.15
grpcio==1.21.1
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
httplib2==0.13.1
idna==2.8
imageio==2.5.0
imagesize==1.1.0
imgaug==0.3.0
importlib-metadata==0.0.0
ipykernel==5.1.1
ipyparallel==6.2.4
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.16
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.3
jeepney==0.4
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyter-http-over-ws==0.0.6
jupyterlab==0.35.4
jupyterlab-server==0.2.0
jws==0.1.3
kafka-python==1.4.6
kaggle==1.5.4
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keylogger==2.7.3
keyring==18.0.0
kiwisolver==1.0.1
kombu==4.6.3
lazy-object-proxy==1.3.1
libarchive-c==2.8
lief==0.9.0
llvmlite==0.28.0
locket==0.2.0
lrutils==0.1.4
lxml==4.3.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
menpo==0.9.2
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
monotonic==1.5
more-itertools==6.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
multiprocess==0.70.8
Naked==0.1.31
navigator-updater==0.2.1
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
nltk==3.4
nose==1.3.7
notebook==5.2.2
numba==0.43.1
numexpr==2.6.9
numpy==1.16.2
numpydoc==0.8.0
oauth2client==3.0.0
olefile==0.46
opencv-contrib-python==4.1.1.26
opencv-python==4.1.0.25
opencv-python-headless==4.1.1.26
openpyxl==2.6.1
packaging==19.0
pandas==0.24.2
pandas-datareader==0.7.4
pandocfilters==1.4.2
parse==1.12.1
parso==0.3.4
partd==0.3.10
passlib==1.6.2
path.py==11.5.0
pathlib2==2.3.3
pathos==0.2.4
patsy==0.5.1
pbkdf2==1.3
pep517==0.6.0
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pkginfo==1.5.0.1
plotly==4.1.1
pluggy==0.9.0
ply==3.11
portpicker==1.2.0
pox==0.2.6
ppft==1.6.6.1
prometheus-client==0.6.0
prompt-toolkit==1.0.16
protobuf==3.8.0
psutil==5.6.1
ptyprocess==0.6.0
py==1.8.0
py-bcrypt==0.4
py4j==0.10.7
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycocotools==2.0.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.4.3
pycurl==7.43.0.2
pydip==0.1.8
pydot==1.4.1
pyee==6.0.0
pyflakes==2.1.1
Pygments==2.3.1
pylint==2.3.1
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyppeteer==0.0.25
pyquery==1.4.0
Pyrebase==3.0.27
pyrsistent==0.14.11
PySocks==1.6.8
pyspark==2.4.3
pytest==4.3.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
python-engineio==3.9.3
python-for-android==2019.8.9
python-jwt==2.0.1
python-slugify==3.0.2
python-socketio==4.3.1
pytoml==0.1.21
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.4.3
QtPy==1.7.0
rake-nltk==1.0.4
redis==3.3.8
requests==2.22.0
requests-html==0.10.0
requests-toolbelt==0.7.0
retrying==1.3.3
rope==0.12.0
rsa==4.0
ruamel-yaml==0.15.46
s3transfer==0.2.1
scikit-image==0.14.2
scikit-learn==0.21.2
scipy==1.2.1
seaborn==0.9.0
SecretStorage==3.1.1
selenium==3.141.0
Send2Trash==1.5.0
sh==1.12.14
Shapely==1.6.4.post2
shellescape==3.4.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
smart-open==1.8.4
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
SQLAlchemy==1.3.1
sqlparse==0.3.0
statsmodels==0.9.0
sympy==1.3
tables==3.5.1
tblib==1.3.2
tensorboard==1.14.0
tensorboardX==1.8
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-hub==0.5.0
tensorflow-serving-api==1.14.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
text-unidecode==1.2
toolz==0.9.0
topia.termextract==1.1.0
torch==1.2.0
torchfile==0.1.0
torchvision==0.4.0
tornado==4.5.3
tqdm==4.31.1
traitlets==4.3.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
virtualenv==16.7.4
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==8.0.2
Werkzeug==0.15.6
widgetsnbextension==3.4.2
wrapt==1.11.1
wurlitzer==1.0.2
xlib==0.21
xlrd==1.2.0
XlsxWriter==1.1.5
xlwt==1.3.0
yacs==0.1.6
yahoo-fin==0.8.2
zict==0.1.4
zipp==0.3.3
zope.interface==4.6.0
